Reduce draw calls for tiles by using larger sprites

+++ source/gfx/color.h
@@ -10,14 +10,14 @@ public:
     return (get(d) << 24) + (get(c) << 16) + (get(b) << 8) + (get(a));
   }
 
-  static inline constexpr int get(int d)
+  static inline constexpr uint8_t get(int d)
   {
     if (d < 0)
       return 255;
 
-    int r = d / 100 % 10;
-    int g = d / 10 % 10;
-    int b = d % 10;
+    uint8_t r = d / 100 % 10;
+    uint8_t g = d / 10 % 10;
+    uint8_t b = d % 10;
 
     return r * 36 + g * 6 + b;
   }
